requirements.txt
# Minimal local requirements (you may pin versions as needed)
pandas
openpyxl
PyYAML
boto3
pyspark
delta-spark
click
